Lootable is moved
