+++ Assets/Scripts/Lootables/Attack.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using DeCrawl.Primitives;
 
 public class Attack : Lootable
 {
