+++ Assets/Scripts/Lootables/Canister.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using DeCrawl.Primitives;
 
 public enum CanisterType { XP, Health }
