+++ Assets/Scripts/Lootables/PlayerLevel.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using DeCrawl.Primitives;
 
 public class PlayerLevel : Lootable
 {
