+++ Assets/Scripts/Lootables/Repair.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using DeCrawl.Primitives;
 
 public class Repair : Lootable
 {
