+++ Assets/Scripts/Player/InventoryRack.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System.Linq;
+using DeCrawl.Primitives;
 
 [RequireComponent(typeof(Lootable))]
 public abstract class InventoryRack : MonoBehaviour
