+++ Assets/Scripts/Primitives/AutoBootLooter.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System.Linq;
+using DeCrawl.Primitives;
 
 public class AutoBootLooter : MonoBehaviour
 {
